Move rpmcanon’s canonicalization code to the library
rpm-ostree needs it.
